Use Listener instead of Visitor
Because I need a list of objects created in DDL for C++ code generation,
it suffices to populate a list of objects as the ANTLR listener encounters
them and modify/add the object's properties as they are encountered (depth-first,
so I only have to add to / modify the last object). The traversal is already
implemented with an ANTLR Listener, so it is preferrable over a Visitor for this
use case.